Reuse PyUrl once
